feat: added xml resources
