+++ apps/mobile_app/src/main/kotlin/dev/marlonlom/apps/talculator/MainActivity.kt
