+++ apps/mobile_app/src/main/res/xml/data_extraction_rules.xml
@@ -4,16 +4,16 @@
    for details.
 -->
 <data-extraction-rules>
-    <cloud-backup>
-        <!-- TODO: Use <include> and <exclude> to control what is backed up.
-        <include .../>
-        <exclude .../>
-        -->
-    </cloud-backup>
-    <!--
-    <device-transfer>
-        <include .../>
-        <exclude .../>
-    </device-transfer>
+  <cloud-backup>
+    <!-- TODO: Use <include> and <exclude> to control what is backed up.
+    <include .../>
+    <exclude .../>
     -->
+  </cloud-backup>
+  <!--
+  <device-transfer>
+      <include .../>
+      <exclude .../>
+  </device-transfer>
+  -->
 </data-extraction-rules>
